+++ cli/src/__tests__/commands/check/check-service.test.ts
